move builder.py to api
