+++ mlrun/api/utils/builder.py
@@ -14,7 +14,6 @@
 import os.path
 import pathlib
 import re
-import tarfile
 import tempfile
 from base64 import b64decode, b64encode
 from os import path
@@ -26,10 +25,13 @@ import mlrun.api.utils.singletons.k8s
 import mlrun.common.schemas
 import mlrun.errors
 import mlrun.runtimes.utils
-
-from .config import config
-from .datastore import store_manager
-from .utils import enrich_image_url, get_parsed_docker_registry, logger, normalize_name
+from mlrun.config import config
+from mlrun.utils import (
+    enrich_image_url,
+    get_parsed_docker_registry,
+    logger,
+    normalize_name,
+)
 
 IMAGE_NAME_ENRICH_REGISTRY_PREFIX = "."
 
@@ -301,17 +303,6 @@ def configure_kaniko_ecr_init_container(
         env=init_container_env,
         name="create-repos",
     )
-
-
-def upload_tarball(source_dir, target, secrets=None):
-
-    # will delete the temp file
-    with tempfile.NamedTemporaryFile(suffix=".tar.gz") as temp_fh:
-        with tarfile.open(mode="w:gz", fileobj=temp_fh) as tar:
-            tar.add(source_dir, arcname="")
-        stores = store_manager.set(secrets)
-        datastore, subpath = stores.get_or_create_store(target)
-        datastore.upload(subpath, temp_fh.name)
 
 
 def build_image(
