+++ mlrun/runtimes/kubejob.py
@@ -22,7 +22,6 @@ import mlrun.common.schemas
 import mlrun.errors
 from mlrun.runtimes.base import BaseRuntimeHandler
 
-from ..builder import build_runtime
 from ..db import RunDBError
 from ..errors import err_to_str
 from ..kfpops import build_op
@@ -215,6 +214,7 @@ class KubejobRuntime(KubeResource):
         if is_kfp:
             watch = True
 
+        ready = False
         if self._is_remote_api():
             db = self._get_db()
             data = db.remote_builder(
@@ -240,17 +240,6 @@ class KubejobRuntime(KubeResource):
                 state = self._build_watch(watch, show_on_failure=show_on_failure)
                 ready = state == "ready"
                 self.status.state = state
-        else:
-            self.save(versioned=False)
-            ready = build_runtime(
-                mlrun.common.schemas.AuthInfo(),
-                self,
-                with_mlrun,
-                mlrun_version_specifier,
-                skip_deployed,
-                watch,
-            )
-            self.save(versioned=False)
 
         if watch and not ready:
             raise mlrun.errors.MLRunRuntimeError("Deploy failed")
